Refactored util filenames and added raw translations csv under data
