+++ src/lib/PrintButton.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type {TranslatedText} from "../utils";
+  import type {TranslatedText} from "./utils/translate-text";
 
   interface PrintButtonProps {
     inputText: string;
