+++ src/lib/utils/english-dovahkul-dictionary.ts
@@ -1,3 +1,9 @@
+// need to test the translator handles a : b
+// a : b,c,d
+// come: meyz
+// come together: grind
+
+// TODO: to be completed from https://www.thuum.org/library/Dragon_Language_Print_Dictionary_5th_Edition.pdf
 export const TRANSLATION_DICTIONARY: {[key: string]: string} = {
     guide: "aak",
     guidance: "aak",
@@ -8,5 +14,5 @@ export const TRANSLATION_DICTIONARY: {[key: string]: string} = {
     possibility: "aal",
     wish: "aal",
     prayer: "aal"
-    // TODO: to be completed from https://www.thuum.org/library/Dragon_Language_Print_Dictionary_5th_Edition.pdf
+
 }
